third_party/blink/web_tests/fast/frames/detached-shadow-frame-expected.txt

CONSOLE WARNING: line 15: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Should not be able to create an iframe with a loaded contentDocument that is not in the document tree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS innerFrame.contentDocument is null
PASS successfullyParsed is true

TEST COMPLETE
